% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/course.R
\name{assignment}
\alias{assignment}
\title{Custom RMarkdown document type that parses student assignments}
\usage{
assignment(...)
}
\arguments{
\item{...}{}
}
\description{
Flag an Rmd file as an assignment by setting its output type to this
function; not meant to be called directly.
}
