% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/course.R
\name{website}
\alias{website}
\title{Open course website}
\usage{
website()
}
\description{
This convenience function will open the course website in the RStudio Viewer
panel.  Only works if run from within RStudio.
}

